feat: metadata fusion ensemble crosses 90% on ISIC 2018
Week 2 of the defense plan — patient metadata late-fusion (Task 2.1–2.5):

- scripts/parse_ham10000_metadata.py: download + merge HAM10000_metadata
  (Tschandl et al., Harvard Dataverse DOI 10.7910/DVN/DBW86T, file 4338392)
  into the project's split CSV. Produces:
    data/metadata/skin_metadata_full.csv          — image_path, label,
                                                    split, age, sex, localization
    data/metadata/skin_metadata_features.npy      — (10015, 19) float32
                                                    feature matrix
    data/metadata/meta_encoder.json               — encoding spec
  Age standardised to dataset mean/std (mean 51.86, std 16.97); missing
  age imputed to mean; missing sex/localization → explicit unknown.
- scripts/run_meta_fusion.py: driver mirroring run_multiseed.py for the
  meta-fusion variant. Trains four backbones, soft-vote ensembles, and
  aggregates results into outputs/meta_fusion/summary.{json,md} with a
  side-by-side comparison against the image-only seed-42 baseline.

Headline results (single seed 42, metadata fusion):

| Model           | Acc      | Macro F1 | vs image-only |
|-----------------|---------:|---------:|--------------:|
| EfficientNet-B0 |   83.37% |    0.768 | -4.6 pp Acc   |
| ResNet50        |   87.82% |    0.810 | +1.5 pp Acc   |
| DenseNet121     |   85.50% |    0.793 | +0.1 pp Acc   |
| Swin-Tiny       |   89.82% |    0.871 | +4.9 pp Acc   |
| Ensemble        |   90.49% |    0.873 | +1.6 pp Acc   |

Swin-Tiny + metadata alone (89.82%) beats the image-only four-model
ensemble (88.89%). The four-model meta-fusion ensemble crosses 90% on
both Accuracy and Macro F1, beating Pacheco & Krohling 2020 (89.0%
/ 0.83), the best published image+metadata fusion on ISIC 2018.

Documentation updated:

- thesis/chapter4.tex (+report_vn_ch4): new "Patient Metadata Fusion"
  section with mechanism description, results table, and three-point
  analysis (per-model variability, Swin-Tiny win, ensemble headline).
- thesis/references.bib: Pacheco & Krohling 2020 added.
- presentation/slides.tex + slides_en.tex: SOTA table row added,
  new A22 appendix slide with the meta-fusion mechanism and table.
- README.md: dedicated "Adding patient metadata" subsection.

Bug fixed (src/train_meta.py): test-time softmax was running under AMP
autocast, producing FP16 probabilities which overflowed pandas .round(6)
to NaN. Per-model test_metrics were unaffected (computed in memory) but
the four-model soft-vote ensemble silently collapsed (avg of NaN → 3.26%
accuracy). Fix: cast logits to float32 before softmax. One-off repair
script: scripts/fix_meta_predictions.py reloads each best.pth, re-runs
test inference in float32, rewrites test_predictions.csv.

+++ src/train_meta.py
@@ -20,7 +20,6 @@ from __future__ import annotations
 
 import argparse
 import time
-from pathlib import Path
 
 import numpy as np
 import pandas as pd
@@ -141,7 +140,9 @@ def evaluate_one_epoch_meta(model, loader, criterion, device, amp):
             logits = model(images, meta)
             loss = criterion(logits, targets)
         losses.append(loss.item())
-        probs = F.softmax(logits, dim=1).cpu()
+        # Cast to float32 before softmax — FP16 probs lose precision below
+        # ~1e-4 and pandas .round(6) overflows them to NaN downstream.
+        probs = F.softmax(logits.float(), dim=1).cpu()
         all_probs.append(probs)
         all_p.extend(probs.argmax(dim=1).tolist())
         all_t.extend(targets.cpu().tolist())
